fixed lots of things in main method, lots more flexible
