+++ .gitignore
@@ -38,6 +38,7 @@ Source/source/out/*
 Source/source/CMakeFiles/*
 
 
+
 # CMake files
 *.cmake
 CMakeCache.txt
@@ -49,4 +50,8 @@ cmake_install.cmake
 install_manifest.txt
 compile_commands.json
 CTestTestfile.cmake
+
+# Other files
+*.DS_Store
+*.vscode
