+++ Source/source/CMakeLists.txt
@@ -1,31 +1,56 @@
 cmake_minimum_required (VERSION 2.6)
 set (CMAKE_CXX_STANDARD 11)
+set (CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_HOME_DIRECTORY}/../Build)
+set (CMAKE_BINARY_DIR ${CMAKE_HOME_DIRECTORY}/../Build)
 
 project (LORIS-GPS)
-SET ( message include/internal-messaging)
+SET ( OS include/internal-messaging)
 
-include_directories ( include ${message}/message/header ${message}/message/identifiers ${message}/unixdomain/header include/minmea)
+include_directories ( 
+   include 
 
-
+   ${OS}/message/header 
+   ${OS}/message/identifiers 
+   ${OS}/unixdomain/header
+   include/minmea
+)
 
 SET ( unixdomain
 
-    ${message}/unixdomain/cc/UnixDomainStreamSocket.cc 
-    ${message}/unixdomain/cc/UnixDomainStreamSocketClient.cc 
-    ${message}/unixdomain/cc/UnixDomainStreamSocketServer.cc )
+    ${OS}/unixdomain/cc/UnixDomainStreamSocket.cc 
+    ${OS}/unixdomain/cc/UnixDomainStreamSocketClient.cc 
+    ${OS}/unixdomain/cc/UnixDomainStreamSocketServer.cc 
+)
+
 SET ( message 
 
-   ${message}/message/cc/Key.cc 
-   ${message}/message/cc/KeyIntPair.cc 
-   ${message}/message/cc/KeyFloatPair.cc
-   ${message}/message/cc/KeyValuePairContainer.cc 
-   ${message}/message/cc/KeyValuePairContainerInterface.cc 
-   ${message}/message/cc/Message.cc 
-   ${message}/message/cc/MessageBuilder.cc 
-   ${message}/message/cc/MessageBuilderInterface.cc  )
+   ${OS}/message/cc/Key.cc 
+   ${OS}/message/cc/KeyIntPair.cc 
+   ${OS}/message/cc/KeyFloatPair.cc
+   ${OS}/message/cc/KeyValuePairContainer.cc 
+   ${OS}/message/cc/KeyValuePairContainerInterface.cc 
+   ${OS}/message/cc/Message.cc 
+   ${OS}/message/cc/MessageBuilder.cc 
+   ${OS}/message/cc/MessageBuilderInterface.cc
+)
+
+SET ( messaging-service 
+
+   ${OS}/messaging-service/cc/MessagingService.cc 
+   ${OS}/messaging-service/cc/MessageSenderInterface.cc 
+)
+
 set ( sanitycheck
 
     SanityCheck.cpp )
 
-#add_executable( Loris-GPS GPS.cpp include/minmea.c ${unixdomain} ${message} )
-add_executable( Loris-GPS GPS.cpp GPSMessageInterface.cpp include/minmea.c ${message} ${sanitycheck} )
+#add_executable( Loris-GPS GPS.cpp include/minmea.c ${unixdomain} ${message} ${messaging-service})
+add_executable( Loris-GPS 
+
+   GPS.cpp 
+   GPSMessageInterface.cpp 
+   GPSDecode.cpp
+   include/minmea.c 
+   ${message} 
+   ${sanitycheck} 
+)
