+++ Source/source/GPSMessageInterface.cpp
@@ -7,6 +7,7 @@ chrono::system_clock::time_point last_empty_poll = chrono::system_clock::now();
 // note: this class can absolutely be mocked out for unit tests
 bool send_message(gps_data* decoded_data) {
 	MessageBuilder messageBuilder;
+	Identifiers identifiers;
 	messageBuilder.StartMessage();
 	
 
@@ -19,8 +20,8 @@ bool send_message(gps_data* decoded_data) {
 	container.AddKeyValuePair(data_types.altitude, decoded_data -> altitude);
 
 	messageBuilder.SetMessageContents(container);
-	messageBuilder.SetRecipient(1097346);
-	messageBuilder.SetSender(6858902);
+	messageBuilder.SetRecipient(identifiers.gps_repository);
+	messageBuilder.SetSender(identifiers.gps_subsystem);
 	
 
 	Message message = messageBuilder.CompleteMessage();
@@ -28,9 +29,8 @@ bool send_message(gps_data* decoded_data) {
 	char msg[256] = "";
 	message.flatten(msg);
 	
-	//SerializeMessage(&message, msg);
 	//MessageSenderInterface ms(message.GetRecipient());
-	//ms.SendMessage(msg);
+	//ms.SendMessage(message);
 
 	
 	message = Message(msg);
@@ -61,14 +61,14 @@ Message get_message(){
 
 	status_codes codes;
 	Identifiers identifiers;
-	//every 1 minute, send a REQUEST message 
-	if (1 - chrono::duration_cast<chrono::minutes>(chrono::system_clock::now() - last_poll).count() == 0) {
+	//every 30 seconds, send a REQUEST message 
+	if (30 - chrono::duration_cast<chrono::seconds>(chrono::system_clock::now() - last_poll).count() == 0) {
 		cout << "TEST: Sending REQUEST message..." << endl;
 		container.AddKeyValuePair(0, codes.request);
 		last_poll = chrono::system_clock::now();
 	}
-	//every 30 seconds, send nothing
-	else if (30 - chrono::duration_cast<chrono::seconds>(chrono::system_clock::now() - last_empty_poll).count() == 0) {
+	//every 20 seconds, send nothing
+	else if (20 - chrono::duration_cast<chrono::seconds>(chrono::system_clock::now() - last_empty_poll).count() == 0) {
 		cout << "TEST: Sending EMPTY message..." << endl;
 		last_empty_poll = chrono::system_clock::now();
 	}
@@ -81,19 +81,4 @@ Message get_message(){
 	builder.SetRecipient(identifiers.gps_subsystem);
 	builder.SetSender(identifiers.gps_repository);
 	return builder.CompleteMessage();
-}
-
-gps_time decode_time_from_int(unsigned int time_int) {
-    unsigned int hours, minutes, seconds;
-    unsigned int BITS = 32;
-    unsigned int H_BITS = 5;
-    unsigned int M_BITS = 6;
-    unsigned int S_BITS = 6;
-    struct gps_time time = {};
-    
-    time.seconds = time_int << (BITS - S_BITS) >> (BITS - S_BITS);
-    time.minutes = time_int << (BITS - M_BITS - S_BITS) >> (BITS - M_BITS);
-    time.hours = time_int << (BITS - H_BITS - M_BITS - S_BITS) >> (BITS - H_BITS);
-    
-    return time;
 }
