+++ Source/source/include/GPSDataTypes.hpp
