+++ Source/source/include/GPSDecode.hpp
@@ -1,0 +1,6 @@
+#ifndef LORIS_GPS_DECODE
+#define LORIS_GPS_DECODE
+
+#include "GPSDataTypes.hpp"
+gps_time decode_time_from_int(unsigned int time_int);
+#endif
