+++ Source/source/include/GPSMessageInterface.hpp
@@ -1,10 +1,9 @@
 #ifndef LORIS_GPS_MESSAGE_INTERFACE
 #define LORIS_GPS_MESSAGE_INTERFACE
 
-#define INDENT_SPACES "  "
 #include "Message.h"
 #include "MessageBuilder.h"
-#include "GPS_Data_Types.h"
+#include "GPSDataTypes.hpp"
 #include <string>
 #include <iostream>
 #include <math.h>
@@ -17,4 +16,8 @@ using namespace std;
 bool send_message(gps_data* decoded_data);
 Message get_message();
 
+#endif
+
+#ifndef INDENT_SPACES
+#define INDENT_SPACES "  "
 #endif
