+++ Source/source/include/SanityCheck.h
@@ -7,7 +7,7 @@ using namespace std;
 
 // testing-only declarations follow, to be removed once we have GPS hardware interface
 static ifstream *nmea_datafile = NULL;
-static string nmea_filename = "nmea01.txt"; // alter filename as needed
+static string nmea_filename = "nmea02.txt"; // alter filename as needed
 
 bool read_nmea_from_file(string *data);
 void close_nmea_file();
